BackEnd/requirements.txt
Flask
pymongo
python-dotenv
uuid
hashlib
